+++ Changelog.md
@@ -3,6 +3,7 @@ ocmtoc changelog
 
 #### v1.0.2
 - Mark `.reloc` and `.debug` as NX
+- Added `-require_read_only_relocs` argument
 
 #### v1.0.1
 - Added support for the `__RELOC_FIX` marker segment